# out/aibot-api-server/api/openapi.yaml
openapi: 3.0.1
info:
  description: aibot api
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: AIBot open api
  version: 1.0.0
servers:
- url: https://openapi.tanyibot.com/apiOpen/v1
tags:
- description: isv
  name: isv
- description: phone
  name: phone
paths:
  /isv/genSubAccountIsv:
    post:
      description: 通过此接口可生成子账号的ISV账号
      operationId: genSubAccountIsv
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetSubAccountIsvReq'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSubAccountIsvResp'
          description: successful operation
      summary: 通过此接口可生成子账号的ISV账号
      tags:
      - isv
      x-codegen-request-body-name: body
  /phone/getPhoneList:
    get:
      description: 通过接口可以获取所有可用的外呼线路的列表
      operationId: getPhoneList
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPhoneListResp'
          description: successful operation
      summary: 通过接口可以获取所有可用的外呼线路的列表
      tags:
      - phone
  /phone/updatePhoneInfoByTenantPhoneNumberId:
    post:
      description: 通过此接口修改线路的归属地、行业、黑名单，只能修改归属客户自己的线路
      operationId: updatePhoneInfoByTenantPhoneNumberId
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePhoneInfoByTenantPhoneNumberIdReq'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: successful operation
      summary: 通过此接口修改线路的归属地、行业、黑名单，只能修改归属客户自己的线路
      tags:
      - phone
      x-codegen-request-body-name: body
  /phone/updatePhonePriceByTenantPhoneNumberId:
    post:
      description: 修改绑定客户线路的价格
      operationId: updatePhonePriceByTenantPhoneNumberId
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
          description: successful operation
      summary: 修改绑定客户线路的价格
      tags:
      - phone
components:
  schemas:
    GetSubAccountIsvReq:
      description: ""
      example:
        phoneNumber: phoneNumber
        tenantSign: tenantSign
      properties:
        tenantSign:
          type: string
        phoneNumber:
          type: string
      required:
      - phoneNumber
      - tenantSign
      title: GetSubAccountIsvReq
      type: object
    GetSubAccountIsvResp:
      description: ""
      example:
        code: 0
        tenantSign: tenantSign
        requestId: requestId
        appKey: appKey
        appSecret: appSecret
        resultMsg: resultMsg
      properties:
        code:
          description: 响应码
          format: int32
          type: integer
        requestId:
          description: 请求Id
          type: string
        resultMsg:
          description: 响应说明
          type: string
        tenantSign:
          description: 客户签名
          type: string
        appKey:
          type: string
        appSecret:
          type: string
      title: GetSubAccountIsvResp
      type: object
    GetPhoneListResp:
      description: ""
      example:
        deadArea: deadArea
        tenantPhoneNumberId: 6
        phoneType: phoneType
        areaCode: areaCode
        code: 0
        phoneNumber: phoneNumber
        localBillRate: localBillRate
        province: province
        city: city
        otherBillRate: otherBillRate
        phoneName: phoneName
        callOutIndustry: callOutIndustry
      properties:
        code:
          description: 响应码
          format: int32
          type: integer
        tenantPhoneNumberId:
          description: 电话id
          format: int64
          type: integer
        phoneNumber:
          description: 电话号码
          type: string
        phoneName:
          description: 电话号码名称
          type: string
        phoneType:
          description: (MOBILE, "手机"),(LANDLINE, "固话"),(UNFIXED_CALL, "无主叫固话")
          type: string
        localBillRate:
          description: 本地话费（单位：厘）
          type: string
        otherBillRate:
          description: 外地话费（单位：厘）
          type: string
        callOutIndustry:
          description: (FINANCE, "金融"),(OTHER, "其他")
          type: string
        areaCode:
          description: 归属地区号
          type: string
        province:
          description: 归属地省
          type: string
        city:
          description: 归属地市
          type: string
        deadArea:
          description: 无法外呼地区
          type: string
      title: APIResponse
      type: object
    UpdatePhoneInfoByTenantPhoneNumberIdReq:
      description: ""
      example:
        deadArea:
          city: city
          prov: prov
        areaCode: http://www.baidu.com
        province: 浙江
        city: 杭州
        callOutIndustry: FINANCE
      properties:
        areaCode:
          description: 线路归属地
          example: http://www.baidu.com
          type: string
        province:
          description: 线路归属地省
          example: 浙江
          type: string
        city:
          description: 线路归属地市
          example: 杭州
          type: string
        callOutIndustry:
          description: (FINANCE, "金融"),(OTHER, "其他")
          enum:
          - FINANCE
          - OTHER
          type: string
        deadArea:
          $ref: '#/components/schemas/UpdatePhoneInfoByTenantPhoneNumberIdReq_deadArea'
      title: UpdatePhoneInfoByTenantPhoneNumberIdReq
      type: object
    APIResponse:
      description: ""
      example:
        requestId: requestId
        resultMsg: resultMsg
      properties:
        requestId:
          description: 请求id
          type: string
        resultMsg:
          description: 响应说明
          type: string
      title: APIResponse
      type: object
    UpdatePhoneInfoByTenantPhoneNumberIdReq_deadArea:
      description: 线路黑名单地址
      example:
        city: city
        prov: prov
      properties:
        city:
          type: string
        prov:
          type: string