# out/aibot-api-server/src/main/kotlin/org/openapitools/model/GetSubAccountIsvReq.kt
package org.openapitools.model

import java.util.Objects
import com.fasterxml.jackson.annotation.JsonProperty
import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size

/**
 * 
 * @param tenantSign 
 * @param phoneNumber 
 */
data class GetSubAccountIsvReq(

    @get:NotNull 
    @field:JsonProperty("tenantSign") val tenantSign: kotlin.String,

    @get:NotNull 
    @field:JsonProperty("phoneNumber") val phoneNumber: kotlin.String
) {

}